# src/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/create":{"post":{"tags":["User"],"summary":"Create User","operationId":"create_user_users_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"Username is already taken"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["User"],"summary":"Get Me","operationId":"get_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSchema"}}}},"401":{"description":"Incorrect auth data"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh-tokens":{"post":{"tags":["Auth"],"summary":"Refresh Tokens","operationId":"refresh_tokens_auth_refresh_tokens_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSchema"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtRefreshBearer":[]},{"JwtRefreshCookie":[]}]}},"/groups/create":{"post":{"tags":["Group"],"summary":"Create Group","operationId":"create_group_groups_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"400":{"description":"Name is taken"},"401":{"description":"Unauthorized"},"404":{"description":"User not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/groups/my_groups":{"get":{"tags":["Group"],"summary":"Get User Groups","operationId":"get_user_groups_groups_my_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Group"},"type":"array","title":"Response Get User Groups Groups My Groups Get"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/groups/get_group_users":{"get":{"tags":["Group"],"summary":"Get All Users Of Group","operationId":"get_all_users_of_group_groups_get_group_users_get","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/USER_GROUP_relationship"},"title":"Response Get All Users Of Group Groups Get Group Users Get"}}}},"400":{"description":"User not in group"},"401":{"description":"Unauthorized"},"403":{"description":"Access forbidden due to user rights"},"404":{"description":"Group not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/group_info":{"get":{"tags":["Group"],"summary":"Get Group Info","operationId":"get_group_info_groups_group_info_get","parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"404":{"description":"Group not found"},"500":{"description":"Internal Server Error"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/my_group_role":{"get":{"tags":["Group"],"summary":"My Relationship With Group","operationId":"my_relationship_with_group_groups_my_group_role_get","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/USER_GROUP_relationship"}}}},"401":{"description":"Unauthorized"},"404":{"description":"User not in group"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invite_group_links/create":{"post":{"tags":["Invite_group_link"],"summary":"Create Invite Link","operationId":"create_invite_link_invite_group_links_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteLinkCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteLink"}}}},"400":{"description":"User not in group"},"401":{"description":"Unauthorized"},"403":{"description":"Access forbidden due to user rights"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/invite_group_links/created_by_me":{"get":{"tags":["Invite_group_link"],"summary":"Get My Invite Links","operationId":"get_my_invite_links_invite_group_links_created_by_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InviteLink"},"type":"array","title":"Response Get My Invite Links Invite Group Links Created By Me Get"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/invite_group_links/for_group":{"get":{"tags":["Invite_group_link"],"summary":"Get My Group Invite Links","operationId":"get_my_group_invite_links_invite_group_links_for_group_get","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InviteLink"},"title":"Response Get My Group Invite Links Invite Group Links For Group Get"}}}},"400":{"description":"User not in group"},"401":{"description":"Unauthorized"},"403":{"description":"Access forbidden due to user rights"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invite_group_links/delete":{"delete":{"tags":["Invite_group_link"],"summary":"Delete Invite Link","operationId":"delete_invite_link_invite_group_links_delete_delete","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized"},"403":{"description":"Access forbidden due to user rights"},"404":{"description":"Invite link not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invite_group_links/search":{"get":{"tags":["Invite_group_link"],"summary":"Search For Link","operationId":"search_for_link_invite_group_links_search_get","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestrictedInviteLink"}}}},"404":{"description":"Invite link not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invite_group_links/use_link":{"post":{"tags":["Invite_group_link"],"summary":"Use Invite Link","operationId":"use_invite_link_invite_group_links_use_link_post","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"link_id","in":"query","required":true,"schema":{"type":"string","title":"Link Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"User is already in group"},"401":{"description":"Unauthorized"},"404":{"description":"Invite link not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/join_group_invites/by_me":{"get":{"tags":["Join_group_invites"],"summary":"Get Join Group Invites Created By Me","operationId":"get_join_group_invites_created_by_me_join_group_invites_by_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/JoinGroupInvite"},"type":"array","title":"Response Get Join Group Invites Created By Me Join Group Invites By Me Get"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/join_group_invites/for_me":{"get":{"tags":["Join_group_invites"],"summary":"Get Join Group Invites For Me","operationId":"get_join_group_invites_for_me_join_group_invites_for_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/RestrictedJoinGroupInvite"},"type":"array","title":"Response Get Join Group Invites For Me Join Group Invites For Me Get"}}}},"401":{"description":"Unauthorized"}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/join_group_invites/create":{"post":{"tags":["Join_group_invites"],"summary":"Create Join Group Invite","operationId":"create_join_group_invite_join_group_invites_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinGroupInviteCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"User is already in group"},"401":{"description":"Unauthorized"},"403":{"description":"Access forbidden due to user rights"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/join_group_invites/accept":{"post":{"tags":["Join_group_invites"],"summary":"Accept Invite","operationId":"accept_invite_join_group_invites_accept_post","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"invite_id","in":"query","required":true,"schema":{"type":"integer","title":"Invite Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"User is already in group"},"401":{"description":"Unauthorized"},"404":{"description":"Invite not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/join_group_invites/decline":{"post":{"tags":["Join_group_invites"],"summary":"Decline Invite","operationId":"decline_invite_join_group_invites_decline_post","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"invite_id","in":"query","required":true,"schema":{"type":"integer","title":"Invite Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized"},"404":{"description":"Invite not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/file/upload":{"post":{"tags":["File"],"summary":"Upload File","operationId":"upload_file_file_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_file_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/File"}}}},"401":{"description":"Incorrect auth data"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/file/download":{"get":{"tags":["File"],"summary":"Download File","operationId":"download_file_file_download_get","parameters":[{"name":"file_id","in":"query","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized"},"404":{"description":"File not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/poll/create":{"post":{"tags":["Polls"],"summary":"Create Poll","operationId":"create_poll_poll_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PollCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Poll"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/poll/created_by_me":{"get":{"tags":["Polls"],"summary":"Get My Polls","operationId":"get_my_polls_poll_created_by_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Poll"},"type":"array","title":"Response Get My Polls Poll Created By Me Get"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/poll/for_group":{"get":{"tags":["Polls"],"summary":"Get Polls For Group","operationId":"get_polls_for_group_poll_for_group_get","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Poll"},"title":"Response Get Polls For Group Poll For Group Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/poll/info":{"get":{"tags":["Polls"],"summary":"Get Poll Info","operationId":"get_poll_info_poll_info_get","parameters":[{"name":"poll_id","in":"query","required":true,"schema":{"type":"integer","title":"Poll Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestrictedPoll"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/join_poll_invites/create":{"post":{"tags":["Join_poll_invites"],"summary":"Create Join Poll Invite","operationId":"create_join_poll_invite_join_poll_invites_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinPollInviteCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}},"/join_poll_invites/for_my_group":{"get":{"tags":["Join_poll_invites"],"summary":"Get Join Poll Invites For My Group","operationId":"get_join_poll_invites_for_my_group_join_poll_invites_for_my_group_get","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"group_id","in":"query","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JoinPollInvite"},"title":"Response Get Join Poll Invites For My Group Join Poll Invites For My Group Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/join_poll_invites/accept":{"post":{"tags":["Join_poll_invites"],"summary":"Accept Join Poll Invite","operationId":"accept_join_poll_invite_join_poll_invites_accept_post","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"join_group_invite_id","in":"query","required":true,"schema":{"type":"integer","title":"Join Group Invite Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/join_poll_invites/decline":{"post":{"tags":["Join_poll_invites"],"summary":"Decline Join Poll Invite","operationId":"decline_join_poll_invite_join_poll_invites_decline_post","security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}],"parameters":[{"name":"join_group_invite_id","in":"query","required":true,"schema":{"type":"integer","title":"Join Group Invite Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/votes/commit_vote":{"post":{"tags":["Votes"],"summary":"Vote To Poll","operationId":"vote_to_poll_votes_commit_vote_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoteCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JwtAccessBearer":[]},{"JwtAccessCookie":[]}]}}},"components":{"schemas":{"AuthSchema":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["access_token","refresh_token"],"title":"AuthSchema"},"Base_group_roles":{"type":"string","enum":["viewer","reviewer","admin"],"title":"Base_group_roles"},"Body_upload_file_file_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_file_upload_post"},"File":{"properties":{"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"filename":{"type":"string","title":"Filename"},"created_by_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Created By Id"}},"type":"object","required":["id","created_at","filename"],"title":"File"},"Group":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"logo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logo"}},"type":"object","required":["id"],"title":"Group"},"GroupCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"GroupCreate"},"Group_roles":{"type":"string","enum":["viewer","reviewer","admin","owner"],"title":"Group_roles"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InviteLink":{"properties":{"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"expires":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires"},"usage_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Usage Limit"},"role":{"$ref":"#/components/schemas/Base_group_roles"},"group_id":{"type":"integer","title":"Group Id"},"created_by_id":{"type":"integer","title":"Created By Id"}},"type":"object","required":["id","created_at","expires","usage_limit","role","group_id","created_by_id"],"title":"InviteLink"},"InviteLinkCreate":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"usage_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Usage Limit"},"expires":{"anyOf":[{"type":"string","format":"duration"},{"type":"null"}],"title":"Expires"},"role":{"$ref":"#/components/schemas/Base_group_roles"}},"type":"object","required":["group_id","role"],"title":"InviteLinkCreate"},"JoinGroupInvite":{"properties":{"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"role":{"$ref":"#/components/schemas/Base_group_roles"},"group_id":{"type":"integer","title":"Group Id"},"for_whom_id":{"type":"integer","title":"For Whom Id"},"created_by_id":{"type":"integer","title":"Created By Id"}},"type":"object","required":["id","created_at","role","group_id","for_whom_id","created_by_id"],"title":"JoinGroupInvite"},"JoinGroupInviteCreate":{"properties":{"role":{"$ref":"#/components/schemas/Base_group_roles"},"group_id":{"type":"integer","title":"Group Id"},"for_whom_id":{"type":"integer","title":"For Whom Id"}},"type":"object","required":["role","group_id","for_whom_id"],"title":"JoinGroupInviteCreate"},"JoinPollInvite":{"properties":{"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"poll_id":{"type":"integer","title":"Poll Id"},"for_whom_id":{"type":"integer","title":"For Whom Id"},"role":{"$ref":"#/components/schemas/Poll_roles"}},"type":"object","required":["id","created_at","poll_id","for_whom_id","role"],"title":"JoinPollInvite"},"JoinPollInviteCreate":{"properties":{"poll_id":{"type":"integer","title":"Poll Id"},"for_whom_id":{"type":"integer","title":"For Whom Id"},"role":{"$ref":"#/components/schemas/Poll_roles"}},"type":"object","required":["poll_id","for_whom_id","role"],"title":"JoinPollInviteCreate"},"Poll":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"document_id":{"type":"string","title":"Document Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"result_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Id"},"state":{"$ref":"#/components/schemas/Poll_states"},"voted_for":{"type":"integer","title":"Voted For"},"voted_against":{"type":"integer","title":"Voted Against"},"voters_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Voters Limit"}},"type":"object","required":["id","title","document_id","created_at","state","voted_for","voted_against"],"title":"Poll"},"PollCreate":{"properties":{"title":{"type":"string","title":"Title"},"document_id":{"type":"string","title":"Document Id"},"expires":{"anyOf":[{"type":"string","format":"duration"},{"type":"null"}],"title":"Expires"},"voters_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Voters Limit"}},"type":"object","required":["title","document_id"],"title":"PollCreate"},"Poll_roles":{"type":"string","enum":["viewer","voter"],"title":"Poll_roles"},"Poll_states":{"type":"string","enum":["active","frozen"],"title":"Poll_states"},"RestrictedInviteLink":{"properties":{"id":{"type":"string","title":"Id"},"role":{"$ref":"#/components/schemas/Base_group_roles"},"group_id":{"type":"integer","title":"Group Id"}},"type":"object","required":["id","role","group_id"],"title":"RestrictedInviteLink"},"RestrictedJoinGroupInvite":{"properties":{"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"role":{"$ref":"#/components/schemas/Base_group_roles"},"group_id":{"type":"integer","title":"Group Id"},"for_whom_id":{"type":"integer","title":"For Whom Id"},"created_by_id":{"type":"integer","title":"Created By Id"}},"type":"object","required":["id","created_at","role","group_id","for_whom_id","created_by_id"],"title":"RestrictedJoinGroupInvite"},"RestrictedPoll":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"document_id":{"type":"string","title":"Document Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"voters_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Voters Limit"}},"type":"object","required":["id","title","document_id","created_at"],"title":"RestrictedPoll"},"USER_GROUP_relationship":{"properties":{"id":{"type":"integer","title":"Id"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"role":{"$ref":"#/components/schemas/Group_roles"},"group_id":{"type":"integer","title":"Group Id"},"user_id":{"type":"integer","title":"User Id"},"added_by_id":{"type":"integer","title":"Added By Id"}},"type":"object","required":["id","added_at","role","group_id","user_id","added_by_id"],"title":"USER_GROUP_relationship"},"User":{"properties":{"username":{"type":"string","title":"Username"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["username","id"],"title":"User"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserSignin":{"properties":{"password":{"type":"string","title":"Password"},"username":{"type":"string","title":"Username"}},"type":"object","required":["password","username"],"title":"UserSignin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoteCreate":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"poll_id":{"type":"integer","title":"Poll Id"},"accepted":{"type":"boolean","title":"Accepted"}},"type":"object","required":["group_id","poll_id","accepted"],"title":"VoteCreate"}},"securitySchemes":{"JwtAccessBearer":{"type":"http","scheme":"bearer"},"JwtAccessCookie":{"type":"apiKey","in":"cookie","name":"access_token_cookie"},"JwtRefreshBearer":{"type":"http","scheme":"bearer"},"JwtRefreshCookie":{"type":"apiKey","in":"cookie","name":"refresh_token_cookie"}}}}